backend/requirements.txt
joblib
torch
scikit-learn==0.23.0
pyyaml
shap
captum
